+++ src/i18n/I18N.php
@@ -177,11 +177,22 @@ class I18N extends \yii\i18n\I18N
      */
     public function getAppLocaleIds(): array
     {
+        $this->_defineAppLocales();
+        return array_keys($this->_appLocaleIds);
+    }
+
+    /**
+     * Defines the list of supported app locale IDs.
+     *
+     * @return void
+     */
+    private function _defineAppLocales(): void
+    {
         if ($this->_appLocaleIds !== null) {
-            return $this->_appLocaleIds;
+            return;
         }
 
-        $localeIds = [
+        $this->_appLocaleIds = [
             Craft::$app->sourceLanguage => true,
         ];
 
@@ -193,7 +204,7 @@ class I18N extends \yii\i18n\I18N
         }
         while (($subDir = readdir($handle)) !== false) {
             if ($subDir !== '.' && $subDir !== '..' && is_dir($dir . DIRECTORY_SEPARATOR . $subDir)) {
-                $localeIds[$subDir] = true;
+                $this->_appLocaleIds[$subDir] = true;
             }
         }
         closedir($handle);
@@ -202,14 +213,25 @@ class I18N extends \yii\i18n\I18N
         $generalConfig = Craft::$app->getConfig()->getGeneral();
         if (!empty($generalConfig->extraAppLocales)) {
             foreach ($generalConfig->extraAppLocales as $localeId) {
-                $localeIds[$localeId] = true;
+                $this->_appLocaleIds[$localeId] = true;
             }
         }
         if ($generalConfig->defaultCpLanguage) {
-            $localeIds[$generalConfig->defaultCpLanguage] = true;
+            $this->_appLocaleIds[$generalConfig->defaultCpLanguage] = true;
         }
+    }
 
-        return $this->_appLocaleIds = array_keys($localeIds);
+    /**
+     * Returns whether the given locale ID is a supported app locale ID.
+     *
+     * @param string $localeId
+     * @return bool
+     * @since 3.6.0
+     */
+    public function validateAppLocaleId(string $localeId): bool
+    {
+        $this->_defineAppLocales();
+        return isset($this->_appLocaleIds[$localeId]);
     }
 
     // Site Locales